chapter-04-streaming/requirements.txt
kafka-python-ng>=2.2.0
pydantic>=2.7.0
